registarLiquidacion(formaulario y js)

+++ src/main/java/cl/awakelab/sprint06/repository/ILiquidacionRepository.java
@@ -3,5 +3,5 @@ package cl.awakelab.sprint06.repository;
 import cl.awakelab.sprint06.entity.Liquidacion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ILiquidacionRepository extends JpaRepository<Liquidacion,Integer> {
+public interface ILiquidacionRepository extends JpaRepository<Liquidacion,Long> {
 }
